配置mybatis有误，mapper注入失败

+++ src/main/java/com/example/demo/service/ILearnResourceService.java
@@ -5,7 +5,7 @@ import com.example.demo.domain.LearnResource;
 public interface ILearnResourceService {
     int addLearnResource(LearnResource learnResource);
 
-    LearnResource getLearnResourceById(int id);
+    LearnResource getLearnResourceById(long id);
 
     int updateLearnResource(LearnResource learnResource);
